refactor from entity to models
