--- counter/src/main/java/progetto/react/com/counter/entity/Utenti.java
@@ -1,43 +1,0 @@
-package progetto.react.com.counter.entity;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import javax.persistence.*;
-@Entity
-@Table(name = "Utenti") // Use a different table name
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter
-@Setter
-public class Utenti {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false, unique = true)
-    String userName;
-    @Column(nullable = false)
-    String psw;
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public void setUserName(String userName) {
-        this.userName = userName;
-    }
-
-    public void setPsw(String psw) {
-        this.psw = psw;
-    }
-}
